2024: Reorganization part 1
Hugo and PaperMod have changed significantly in the past two years. A
fresh git clone of this blog no longer builds when run with `hugo
server`.

I'm trying to make this work again, and the first step (I think) was to
clean up the structure of posts. I did this by moving all posts from the
previous <yyyy>/<mm>/<dd>/<post-name.md> to
/content/<yyyy-mm-dd-post-name.md>. I expect this will make referencing
posts easier.

Two years ago, if I wanted a banner icon for a post, it had to be
included as a specially-named image in the post's directory. My
assumption is that I will be able to replicate this functionality - if
desired - using the post's frontmatter.

Fingers crossesd that this commit doesn't screw everything up more.